chore: Better files org + added metric type and config id to tracking ++
