+++ src/GmdVizPanel/config/panel-heights.ts
@@ -1,7 +1,7 @@
 // height in pixels
 export enum PANEL_HEIGHT {
-    S = 160,
-    M = 220,
-    L = 260,
-    XL = 280,
-  }
+  S = 160,
+  M = 220,
+  L = 260,
+  XL = 280,
+}
