+++ src/GmdVizPanel/getMetricType.ts
@@ -1,0 +1,26 @@
+import { isHistogramMetric } from './heatmap/isHistogramMetric';
+import { isAgeMetric } from './isAgeMetric';
+import { isRateQuery } from './isRateQuery';
+import { isUpDownMetric } from './statushistory/isUpDownMetric';
+
+export type MetricType = 'status-updown' | 'histogram' | 'age' | 'rate' | 'non-rate';
+
+export function getMetricType(metric: string): MetricType {
+  if (isUpDownMetric(metric)) {
+    return 'status-updown';
+  }
+
+  if (isHistogramMetric(metric)) {
+    return 'histogram';
+  }
+
+  if (isAgeMetric(metric)) {
+    return 'age';
+  }
+
+  if (isRateQuery(metric)) {
+    return 'rate';
+  }
+
+  return 'non-rate';
+}
